Add logic

+++ app/Services/PostService.php
@@ -2,7 +2,9 @@
 
 namespace App\Services;
 
+use App\Exceptions\PermissionException;
 use App\Repositories\PostRepository;
+
 
 class PostService
 {
@@ -19,7 +21,7 @@ class PostService
 
     public function getLastPosts()
     {
-        return $this->repository->limit(10);
+        return $this->repository->limit(10)->get();
     }
 
     public function create(array $post)
@@ -33,8 +35,13 @@ class PostService
 
     public function update(array $post, $postId)
     {
+        $authorId = $this->repository->findByField('id',$postId)->value('user_id');
+
+        if(auth()->user()->id != $authorId){
+            throw new PermissionException;
+        }
+
         return $this->repository->update([
-            'user_id'   => auth()->user()->id,
             'title'     => $post['title'],
             'content'   => $post['content']
         ], $postId);
@@ -42,11 +49,20 @@ class PostService
 
     public function delete($postId)
     {
+        $authorId = $this->repository->findByField('id',$postId)->value('user_id');
+
+        if(auth()->user()->id != $authorId){
+            throw new PermissionException;
+        }
+
         return $this->repository->delete($postId);
     }
 
     public function getMyPosts()
     {
-        return $this->repository->where('user_id', auth()->user()->id)->limit(10);
+        return $this->repository
+            ->where('user_id', auth()->user()->id)
+            ->limit(10)
+            ->get();
     }
 }
